all elements in locked folders are locked now

+++ src/MovieSelection.py
@@ -1605,6 +1605,7 @@ class EMCSelection(Screen, HelpableScreen, SelectionEventInfo, VlcPluginInterfac
 
 	def deleteMovieConfimation(self, confirmed):
 		current = self.getCurrent()
+		pathname = os.path.dirname(current.getPath())
 		if confirmed and self.tmpSelList is not None and len(self.tmpSelList)>0:
 			if self.delCurrentlyPlaying:
 				if self.playerInstance is not None:
@@ -1612,7 +1613,10 @@ class EMCSelection(Screen, HelpableScreen, SelectionEventInfo, VlcPluginInterfac
 			delete = not config.EMC.movie_trashcan_enable.value or self.permanentDel
 			if os.path.exists(config.EMC.movie_trashcan_path.value) or delete:
 				# if the user doesn't want to keep the movies in the trash, purge immediately
-				self.execFileOp(config.EMC.movie_trashcan_path.value, current, self.tmpSelList, op="delete", purgeTrash=delete)
+				if not self.isUpperPathLocked(pathname):
+					self.execFileOp(config.EMC.movie_trashcan_path.value, current, self.tmpSelList, op="delete", purgeTrash=delete)
+				else:
+					self.session.open(MessageBox, _("This element belongs to a locked folder and cannot be deleted!"), MessageBox.TYPE_ERROR, 10)
 				
 				for x in self.tmpSelList:
 					self.lastPlayedCheck(x)
@@ -1621,6 +1625,15 @@ class EMCSelection(Screen, HelpableScreen, SelectionEventInfo, VlcPluginInterfac
 				self.session.openWithCallback(self.trashcanCreate, MessageBox, _("Delete failed because the trashcan directory does not exist. Attempt to create it now?"), MessageBox.TYPE_YESNO)
 			emcDebugOut("[EMCMS] deleteMovie")
 
+	def isUpperPathLocked(self, path):
+		locked = False
+		while path != "/":
+			if os.path.isfile(path + "/dir.lock"):
+				locked = True
+				break
+			(path, tail) = os.path.split(path)
+		return locked
+	
 	def isPathLocked(self, path):
 		locked = False
 		for root, dirs, files in os.walk(path):
@@ -1630,14 +1643,13 @@ class EMCSelection(Screen, HelpableScreen, SelectionEventInfo, VlcPluginInterfac
 		return locked
 	
 	def delPathSelRecursive(self, service, path, confirm):
-		if confirm:
-			if not self.isPathLocked(path):
-				if path and service:
-					emcTasker.shellExecute('rm -rf "' + path + '"')
-					self.removeService(service)
-					self.setReturnCursor()
-			else:
-				self.session.open(MessageBox, _("The folder or a contained folder is locked, unlock it first!"), MessageBox.TYPE_ERROR, 10)
+		if confirm and path:
+			if self.isPathLocked(path) or self.isUpperPathLocked(path):
+				self.session.open(MessageBox, _("This folder is locked or is in relation to a locked folder, unlock it first!"), MessageBox.TYPE_ERROR, 10)
+			elif service:
+				emcTasker.shellExecute('rm -rf "' + path + '"')
+				self.removeService(service)
+				self.setReturnCursor()		
 
 	def delPathSelConfirmed(self, service, confirm):
 		if confirm and service:
